[refactoring] Clean-up for stlink-flash & -info
- Ensure proper function declaration
- Checked & revised header includes
