+++ src/st-flash/flash_opts.c
@@ -1,11 +1,20 @@
+/*
+ * File: flash_opts.c
+ *
+ * Flash Options
+ */
+
 #include <stdint.h>
-#include <stdlib.h>
 #include <stdio.h>
+#include <stdlib.h>
 #include <string.h>
 
-#include <helper.h>
-
+#include <stm32.h>
+#include <stlink.h>
+#include "flash_opts.h"
 #include "flash.h"
+
+#include <helper.h>
 
 static bool starts_with(const char * str, const char * prefix) {
     uint32_t n = strlen(prefix);
